Move deprecated into its own file
