+++ src/blocks/buttons/block.json
@@ -1,0 +1,18 @@
+{
+	"name": "coblocks/buttons",
+	"category": "coblocks",
+	"attributes": {
+		"items": {
+			"type": "number",
+			"default": "2"
+		},
+		"contentAlign": {
+			"type": "string",
+			"default": "left"
+		},
+		"isStackedOnMobile": {
+			"type": "boolean",
+			"default": false
+		}
+	}
+}
